requirements.txt
openpyxl
pandas
fpdf2
imagesize
